Improve styling and vimeo video loading
Added "background-size: cover" to make background images cover the whole area of an element, added a "border-radius to the video box", set "background-position: center", and made fetched vimeo video thumbnail to act as load screen before the video is loaded. Typewriter content and options were also updated to simplify the text and remove the loop. The changes were made to improve user experience and website performance.